logic bug near wget; always add WRPASS

+++ automated_macOS_install.sh
@@ -77,7 +77,7 @@ printf "${WHITE}preparing the installation files and directories${NC}\n"
 #done
 
 if [ ! -d $HIPPHOME ]; then
-    /bin/mkdir $HIPPHOME
+  /bin/mkdir $HIPPHOME
 fi
 
 cd $HIPPHOME
@@ -128,6 +128,7 @@ else
   echo "$(/usr/local/bin/psql -V) installed; will not ask brew to install psql"
 fi
 
+if [ ! -f '/usr/local/bin/wget' ]; then
   $BREW install wget
 else
   echo "wget already installed; will not ask brew to install wget"
@@ -207,10 +208,8 @@ if [ ! -d "$SERVERPATH/settings/" ]; then
   CONFIGFILE="$SERVERPATH/server/settings/securitysettings.py"
   sed -i "" "s/DBPASS = 'yourpassheretrytomakeitstrongplease'/DBPASS = '$RDPASS'/" $CONFIGFILE
   sed -i "" "s/SECRET_KEY = 'yourkeyhereitshouldbelongandlooklikecryptographicgobbledygook'/SECRET_KEY = '$SKRKEY'/" $CONFIGFILE
-  if [ "$VECTORS" == "y" ]; then
-    sed -i "" "s/WRITEUSER = 'consider_re-using_HipparchiaBuilder_user'/WRITEUSER = 'hippa_wr'/" $CONFIGFILE
-    sed -i "" "s/DBWRITEPASS = 'consider_re-using_HipparchiaBuilder_pass'/DBWRITEPASS = '$WRPASS'/" $CONFIGFILE
-  fi
+  sed -i "" "s/WRITEUSER = 'consider_re-using_HipparchiaBuilder_user'/WRITEUSER = 'hippa_wr'/" $CONFIGFILE
+  sed -i "" "s/DBWRITEPASS = 'consider_re-using_HipparchiaBuilder_pass'/DBWRITEPASS = '$WRPASS'/" $CONFIGFILE
   sed -i "" "s/DEFAULTREMOTEPASS = 'yourremoteuserpassheretrytomakeitstrongplease'/DEFAULTREMOTEPASS = '$RUPASS'/" $CONFIGFILE
   # note: this only works if pg_hba.conf has 'trust' in localhost for `whoami`
   /usr/local/bin/psql -d $THEDB --command="ALTER ROLE hippa_rd WITH PASSWORD '$RDPASS';"
